requirement.txt
asgiref==3.3.4
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
